requirements.txt
# Azure OpenAI Realtime API dependencies
azure-identity==1.25.1
python-dotenv==1.2.1

# Web server
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.0

# Azure Application Insights / OpenTelemetry
azure-monitor-opentelemetry==1.6.0
opentelemetry-instrumentation-fastapi==0.47b0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-playwright==0.5.2
httpx==0.27.0

# Local development only (not needed in container)
# PyAudio==0.2.14
